refactor: reword "bot" to "strategy" in core codebase build: reword "bot" to "liquidator" in Docker-related files
