+++ .env.example
@@ -7,7 +7,7 @@ ALCHEMY_KEY=xxxxxxxxxxxxxxxxxx
 # Infura API Key
 INFURA_KEY=xxxxxxxxxxxxxxxxxx
 
-# Network where the bot should run
+# Network where the selected strategy should run
 NETWORK_NAME=homestead
 
 # Persistence is whether blockchain data should be cached locally across sessions.
