+++ deployment.yaml
@@ -69,7 +69,7 @@ spec:
               cpu: "0.01"
               memory: 64Mi
           volumeMounts:
-            - mountPath: /bot/cache
+            - mountPath: /liquidator/cache
               name: cache
       restartPolicy: Always
       volumes:
