+++ src/index.ts
@@ -1,7 +1,7 @@
-import { Bot } from "./bot/index";
 import { getFlashbotsURL, isTrueSet } from "./helpers";
 import * as networkConfig from "./network-config.json";
-import { NetworkName, Strategy } from "./types";
+import { Strategy as UniswapV2Strategy } from "./strategies/uniswap-v2";
+import { NetworkName, StrategyName } from "./types";
 import { Wallet, providers, utils } from "ethers";
 
 require("dotenv").config();
@@ -10,7 +10,10 @@ const { ALCHEMY_KEY, INFURA_KEY, PERSISTENCE_ENABLED, SELECTED_ACCOUNT, WALLET_S
   [key: string]: string;
 };
 
-const { NETWORK_NAME, SELECTED_STRATEGY } = process.env as { NETWORK_NAME: NetworkName; SELECTED_STRATEGY: Strategy };
+const { NETWORK_NAME, SELECTED_STRATEGY } = process.env as {
+  NETWORK_NAME: NetworkName;
+  SELECTED_STRATEGY: StrategyName;
+};
 
 const account = utils.HDNode.fromMnemonic(WALLET_SEED as string).derivePath(`m/44'/60'/0'/0/${SELECTED_ACCOUNT}`);
 
@@ -23,7 +26,7 @@ const provider = networkConfig[NETWORK_NAME].flashbotsEnabled
 
 const signer = new Wallet(account, provider);
 
-new Bot({
+new UniswapV2Strategy({
   networkConfig: networkConfig[NETWORK_NAME],
   persistenceEnabled: isTrueSet(PERSISTENCE_ENABLED),
   provider,
